move ibft status metrics to ibft package
