+++ validator/validator.go
@@ -102,12 +102,8 @@ func (v *Validator) Start() error {
 
 		for _, ib := range v.ibfts { // init all ibfts
 			go func(ib ibft.Controller) {
-				ReportIBFTStatus(v.Share.PublicKey.SerializeToHexStr(), false, false)
 				if err := ib.Init(); err != nil {
 					v.logger.Error("could not initialize ibft instance", zap.Error(err))
-					ReportIBFTStatus(v.Share.PublicKey.SerializeToHexStr(), false, true)
-				} else {
-					ReportIBFTStatus(v.Share.PublicKey.SerializeToHexStr(), true, false)
 				}
 			}(ib)
 		}
